Basic core structure complete

+++ gr-utils/python/blocktool/core/parseheader.py
@@ -18,7 +18,7 @@
 # the Free Software Foundation, Inc., 51 Franklin Street,
 # Boston, MA 02110-1301, USA.
 #
-""" Module to generate AST for the headers """
+""" Module to generate AST for the headers and parse it """
 
 from __future__ import print_function
 from __future__ import absolute_import
@@ -28,11 +28,15 @@ import os
 import sys
 import logging
 
+import click
+
 from pygccxml import parser
 from pygccxml import declarations
 from pygccxml import utils
 
 from blocktool.core.base import BlockToolException, BlockTool
+from blocktool.core.yaml_generator import yaml_generator
+from blocktool.core.json_generator import json_generator
 
 logger = logging.getLogger(__name__)
 
@@ -51,12 +55,12 @@ class BlockToolGenerateAst(BlockTool):
             module_types.append(list_dir)
     module_types.remove('utils')
     module_types = tuple(module_types)
-    
 
     def __init__(self, module_name=None, file_name=None, **kwargs):
         """ __init__ """
         BlockTool.__init__(self, **kwargs)
         self.info['modname'] = module_name
+        self.info['filename'] = file_name
 
     def validate(self):
         """ Validates the arguments """
@@ -68,22 +72,44 @@ class BlockToolGenerateAst(BlockTool):
         Abstract Syntax Tree text file generator
         magic happens here!
         """
-        generator_path, generator_name = utils.find_xml_generator()
-        xml_generator_config = parser.xml_generator_configuration_t(
-            xml_generator_path=generator_path,
-            xml_generator=generator_name,
-            compiler="gcc")
-        gr = self.info['modname'].split("-")[0]
-        module = self.info['modname'].split("-")[-1]
-        decls = parser.parse([self.info['target_file']], xml_generator_config)
-        global_namespace = declarations.get_global_namespace(decls)
-        ns = global_namespace.namespace(gr)
-        main_namespace = ns.namespace(module)
-        f = open('temp_ast.txt', 'w')
-        sys.stdout = f
-        return declarations.print_declarations(main_namespace)
+        if self.info['yaml_confirm'] or self.info['json_confirm']:
+            generator_path, generator_name = utils.find_xml_generator()
+            xml_generator_config = parser.xml_generator_configuration_t(
+                xml_generator_path=generator_path,
+                xml_generator=generator_name,
+                compiler="gcc")
+            gr = self.info['modname'].split("-")[0]
+            module = self.info['modname'].split("-")[-1]
+            decls = parser.parse(
+                [self.info['target_file']], xml_generator_config)
+            global_namespace = declarations.get_global_namespace(decls)
+            ns = global_namespace.namespace(gr)
+            main_namespace = ns.namespace(module)
+            orig_stdout = sys.stdout
+            f = open('temp_ast.txt', 'w')
+            sys.stdout = f
+            declarations.print_declarations(main_namespace)
+            sys.stdout = orig_stdout
+            f.close()
+        else:
+            click.secho("Won't do the hardwork unnecessarily!!", fg='blue')
+            exit(1)
+
+    def parse_ast(self):
+        """
+        Parse the AST generated
+        """
+        # Create a data dict after parsing the text file
+        # pass the data dict to both yaml and json generator
+        # do delete the temporary ast file generated
+        if self.info['yaml_confirm']:
+            yaml_generator(self.info['yaml_confirm'])
+
+        if self.info['json_confirm']:
+            json_generator(self.info['json_confirm'])
 
     def run(self):
         """ Run, run, run. """
         self.validate()
         self.ast_generator()
+        self.parse_ast()
